improvement: remove float64 casting in contains method

+++ range.go
@@ -1,10 +1,14 @@
 package main
 
+// Range defined integer value range
 type Range struct {
+	// Min value
 	Min int64
+
+	// Max value
 	Max int64
 }
 
-func (r Range) contains(f float64) bool {
-	return f >= float64(r.Min) && f <= float64(r.Max)
+func (r Range) contains(f int64) bool {
+	return f >= r.Min && f <= r.Max
 }
